webframe: fold util::HttpHeaders into Headers
It was a duplicated type.
