+++ src/util.rs
@@ -19,7 +19,6 @@ use crate::yattag;
 use anyhow::anyhow;
 use anyhow::Context;
 use lazy_static::lazy_static;
-use std::borrow::Cow;
 use std::collections::HashMap;
 use std::collections::HashSet;
 use std::convert::TryFrom;
@@ -1017,21 +1016,6 @@ pub fn get_in_both<T: Clone + Diff>(first: &[T], second: &[T]) -> Vec<T> {
 pub fn get_content(path: &str) -> anyhow::Result<Vec<u8>> {
     // TODO just use std::fs::read() directly, this was 12 lines originally.
     Ok(std::fs::read(path)?)
-}
-
-// TODO move to webframe and avoid the duplication with webframe::Headers.
-type HttpHeaders = Vec<(Cow<'static, str>, Cow<'static, str>)>;
-
-/// Gets the content of a file in workdir with metadata.
-pub fn get_content_with_meta(path: &str) -> anyhow::Result<(Vec<u8>, HttpHeaders)> {
-    let buf = get_content(path)?;
-
-    let metadata = std::fs::metadata(path)?;
-    let modified = metadata.modified()?;
-    let modified_utc: chrono::DateTime<chrono::offset::Utc> = modified.into();
-
-    let extra_headers = vec![("Last-Modified".into(), modified_utc.to_rfc2822().into())];
-    Ok((buf, extra_headers))
 }
 
 /// Determines the normalizer for a given street.
